+++ test/tests_vector.js
@@ -39,7 +39,7 @@ QUnit.module('Vector internal module', {
 
 QUnit.test('Vector basic manipulations', function(assert) {    
   // Test vector creation using random data
-  var vect = PortfolioAllocation.Vector_(this.vectValues);
+  var vect = new PortfolioAllocation.Vector(this.vectValues);
 
   // Check the dimensions	
   assert.equal(vect.nbRows, this.vectValues.length, 'Number of rows');
@@ -57,22 +57,22 @@ QUnit.test('Vector basic manipulations', function(assert) {
 
 QUnit.test('Vector to array', function(assert) {    
   // Test vector to array using static data
-  var vec = PortfolioAllocation.Vector_([1,2,3]);
+  var vec = new PortfolioAllocation.Vector([1,2,3]);
   assert.deepEqual(vec.toArray(), [1,2,3], 'Vector to array');
   
   // Test vector to array using the random vector, using identity vector(vector to array) == vector
-  var nsVec = PortfolioAllocation.Vector_(this.vectValues);
+  var nsVec = new PortfolioAllocation.Vector(this.vectValues);
   var nsVecArray = nsVec.toArray();
   assert.deepEqual(nsVecArray, this.vectValues, 'Array to vector to array');
-  assert.equal(PortfolioAllocation.matrixIdentical_(nsVec, PortfolioAllocation.Vector_(nsVecArray)), true, 'Vector to array to vector');
+  assert.equal(PortfolioAllocation.Vector.areEqual(nsVec, new PortfolioAllocation.Vector(nsVecArray)), true, 'Vector to array to vector');
 });
 
 
 QUnit.test('Vector dot product', function(assert) {    
   // Test vector dot product using static data
-  var vec1 = PortfolioAllocation.Vector_([1,2,3]);
-  var vec2 = PortfolioAllocation.Vector_([4,5,6]);
-  assert.equal(PortfolioAllocation.vectorDotProduct_(vec1, vec2), 32, 'Vector dot product');
+  var vec1 = new PortfolioAllocation.Vector([1,2,3]);
+  var vec2 = new PortfolioAllocation.Vector([4,5,6]);
+  assert.equal(PortfolioAllocation.Vector.dotProduct(vec1, vec2), 32, 'Vector dot product');
   
   // Test vector dot product using formula n(n+1)(2n+1)/6 == dot([1,2,3,...,n],[1,2,3...,n])
   var n_max = Math.floor(Math.random()*(100-2+1) + 2); // min = 2, max = 100
@@ -80,39 +80,52 @@ QUnit.test('Vector dot product', function(assert) {
   for (var n = 1; n <= n_max; ++n) {
     intArray[n-1] = n;
   }
-  var vec = PortfolioAllocation.Vector_(intArray);
-  assert.equal(PortfolioAllocation.vectorDotProduct_(vec, vec), n_max*(n_max+1)*(2*n_max+1)/6, 'Vector dot product integer');
+  var vec = new PortfolioAllocation.Vector(intArray);
+  assert.equal(PortfolioAllocation.Vector.dotProduct(vec, vec), n_max*(n_max+1)*(2*n_max+1)/6, 'Vector dot product integer');
 });
+
+
+QUnit.test('Vector Hadamard product', function(assert) {    
+  // Test vector Hadamard product using static data
+  var vec1 = new PortfolioAllocation.Vector([1,2,3]);
+  var vec2 = new PortfolioAllocation.Vector([1,2,3]);
+  
+  var hadamardVec = PortfolioAllocation.Vector.hadamardProduct(vec1, vec2);
+  var expectedVec = new PortfolioAllocation.Vector([1,4,9]);
+  
+  assert.equal(PortfolioAllocation.Vector.areEqual(hadamardVec, expectedVec), true, 'Vector Hadamard product');
+});
+
 
 
 QUnit.test('Matrix-vector product', function(assert) {    
   // Test matrix-vector product using static data
-  var mat = PortfolioAllocation.Matrix_([[1,2,3], [4,5,6]]);
-  var vec = PortfolioAllocation.Vector_([1,2,3]);
-  var prodVec = PortfolioAllocation.matrixVectorProduct_(mat, vec);
+  var mat = new PortfolioAllocation.Matrix([[1,2,3], [4,5,6]]);
+  var vec = new PortfolioAllocation.Vector([1,2,3]);
+  var prodVec = PortfolioAllocation.Matrix.vectorProduct(mat, vec);
 
-  var expectedResVec = PortfolioAllocation.Vector_([14,32]);
-  assert.equal(PortfolioAllocation.matrixIdentical_(prodVec, expectedResVec), true, 'Matrix-vector product');
+  var expectedResVec = new PortfolioAllocation.Vector([14,32]);
+  assert.equal(PortfolioAllocation.Vector.areEqual(prodVec, expectedResVec), true, 'Matrix-vector product');
 });
 
 
 QUnit.test('Vector sum', function(assert) {    
   // Test vector sum using formula n == sum([1,1,1,...,1])
   var n = Math.floor(Math.random()*(100-2+1) + 2); // min = 2, max = 100
-  var vec = PortfolioAllocation.vectorOnes_(n); 
+  var vec = PortfolioAllocation.Vector.ones(n); 
   assert.equal(vec.sum(), n, 'Vector sum');
 });
 
 
 QUnit.test('Vector normalisation', function(assert) {    
   // Static data
-  var vec = PortfolioAllocation.Vector_([1,2,3]);
-  var resvec = PortfolioAllocation.Vector_([1/6,2/6,3/6]);
-  assert.equal(PortfolioAllocation.matrixIdentical_(vec.normalize(), resvec), true, 'Vector static normalized');
+  var vec = new PortfolioAllocation.Vector([1,2,3]);
+  var resvec = new PortfolioAllocation.Vector([1/6,2/6,3/6]);
+  assert.equal(PortfolioAllocation.Vector.areEqual(vec.normalize(), resvec), true, 'Vector static normalized');
 
   // Vector of ones
   var n = Math.floor(Math.random()*(100-2+1) + 2); // min = 2, max = 100
-  var vec = PortfolioAllocation.vectorOnes_(n).normalize(); 
+  var vec = PortfolioAllocation.Vector.ones(n).normalize(); 
   for (var i = 1; i <= n; ++i) {
     assert.equal(vec.getValueAt(i,1), 1/n, 'Vector of ones normalized');
   }
@@ -122,7 +135,7 @@ QUnit.test('Vector normalisation', function(assert) {
 QUnit.test('Specific vectors creation', function(assert) {    
   // Vector of ones
   var n = Math.floor(Math.random()*(100-2+1) + 2); // min = 2, max = 100
-  var vec = PortfolioAllocation.vectorOnes_(n); 
+  var vec = PortfolioAllocation.Vector.ones(n); 
   
   for (var i = 1; i <= n; ++i) {
     assert.equal(vec.getValueAt(i,1), 1, 'Vector of ones');