+++ changelog.md
@@ -1,3 +1,8 @@
+### 0.0.2 - (on-going)
+
+- New portfolio allocation methods: most diversified portfolio, minimum correlation portfolio, proportional minimum variance portfolio
+- New Matrix/Vector operations, plus code refactoring for future split
+
 ### 0.0.1 - 09/06/2017
 
 - Initial commit